% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers_and_formatters.R
\name{check_wear_time}
\alias{check_wear_time}
\title{Determine if wear time algorithm needs to be run}
\usage{
check_wear_time(df, id, wear)
}
\arguments{
\item{df}{data frame of accelerometer input}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{wear}{character scalar [optional]. Column name of the variable to use
for determining wear time (logical vector with \code{TRUE} for wear time
minutes). If no value is provided, \code{\link{nhanes_wear}} is invoked on
the \code{counts} column}
}
\description{
If needed, the function will run \code{\link{nhanes_wear}} (with a message).
To bypass, run \code{df$is_wear <- TRUE} prior to executing
\code{check_wear_time(df, id, wear = "is_wear")}
}
\keyword{internal}
