% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Data for use in code examples throughout the \code{SBprofiles} package}
\format{
An object of class \code{data.frame} with 10080 rows and 8 columns.
}
\usage{
example_data
}
\description{
This is a minute-by-minute data set for one participant in the National Health
and Nutrition Examination Survey (NHANES). For details, see the
\href{https://wwwn.cdc.gov/Nchs/Nhanes/2003-2004/PAXRAW_C.htm#Codebook}{NHANES documentation}.
}
\keyword{datasets}
