% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bouts.R, R/get_profile.R, R/zzz.R
\name{sb_bout_dist_default}
\alias{sb_bout_dist_default}
\alias{sb_bout_dist_df}
\alias{id_verify}
\alias{counts_verify}
\alias{id_bind}
\alias{internal_functions}
\title{Functions used internally in the package}
\usage{
sb_bout_dist_default(is_sb, is_wear, min_bout, valid_indices, probs)

sb_bout_dist_df(df, counts, id, sb, min_bout, valid_indices, probs)

id_verify(object, id)

counts_verify(object, counts)

id_bind(result, id)
}
\arguments{
\item{is_sb}{logical vector reflecting minute-by-minute classifications
(\code{TRUE} for sedentary behavior and \code{FALSE} for non-sedentary
behavior)}

\item{is_wear}{logical vector reflecting minute-by-minute wear
classifications (\code{TRUE} for wearing and \code{FALSE} for not wearing)}

\item{min_bout}{integer scalar. Threshold for a run to qualify as a bout}

\item{valid_indices}{integer vector (optional) specifying which indices of
\code{is_sb} and {is_wear} correspond to valid measurement days (e.g. with
10+ hours of wear time on 4+ days)}

\item{probs}{numeric vector. Percentiles to calculate (all values must be
between 0 and 1)}

\item{df}{data frame input (for vectorized operations)}

\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{sb}{integer. The cut point to use for classifying sedentary behavior}

\item{result}{an output data frame that may need id-based formatting}
}
\description{
Functions used internally in the package
}
\keyword{internal}
