% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_nonwear.R
\name{nhanes_nonwear}
\alias{nhanes_nonwear}
\title{Retrieve non-wear information for an NHANES accelerometer data file}
\usage{
nhanes_nonwear(counts)
}
\arguments{
\item{counts}{vector of count values (minute-by-minute)}
}
\value{
a vector of wear time information
}
\description{
This is a wrapper around \code{PhysicalActivity::wearingMarking} that takes
into account the non-POSIX formatting of NHANES timestamps
}
\examples{
data(example_data, package = "SBprofiles")
nhanes_nonwear(example_data$PAXINTEN)
}
