% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_wear.R
\name{nhanes_wear}
\alias{nhanes_wear}
\title{Retrieve wear information for an NHANES accelerometer data file}
\usage{
nhanes_wear(counts)
}
\arguments{
\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior (and wear time, depending on the function)}
}
\value{
a vector of wear time information (\code{TRUE} means wearing and
  \code{FALSE} means not wearing)
}
\description{
This is a wrapper around \code{PhysicalActivity::wearingMarking} that takes
into account the non-POSIX formatting of NHANES timestamps
}
\examples{
data(example_data, package = "SBprofiles")
nhanes_wear(example_data$PAXINTEN)
}
