% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bouts.R
\name{sb_bout_dist}
\alias{sb_bout_dist}
\title{Extract features to summarize a participant's bout duration distribution}
\usage{
sb_bout_dist(
  is_sb,
  is_wear,
  min_bout = 5,
  valid_indices = NULL,
  probs = c(0.1, 0.2, 0.25, seq(0.3, 0.7, 0.1), 0.75, 0.8, 0.9)
)
}
\arguments{
\item{is_sb}{logical vector reflecting minute-by-minute classifications
(\code{TRUE} for sedentary behavior and \code{FALSE} for non-sedentary
behavior)}

\item{is_wear}{logical vector reflecting minute-by-minute wear
classifications (\code{TRUE} for wearing and \code{FALSE} for not wearing)}

\item{min_bout}{integer scalar. Threshold for a run to qualify as a bout}

\item{valid_indices}{integer vector (optional) specifying which indices of
\code{is_sb} and {is_wear} correspond to valid measurement days (e.g. with
10+ hours of wear time on 4+ days)}

\item{probs}{numeric vector. Percentiles to calculate (all values must be
between 0 and 1)}
}
\value{
A bout object (data frame of distribution features)
}
\description{
Extract features to summarize a participant's bout duration distribution
}
