% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBprofiles.R
\docType{package}
\name{SBprofiles}
\alias{SBprofiles}
\title{Learn about and use sedentary behavior profiles}
\description{
Package and repository for learning about and using the
  sedentary profiles
}
\seealso{
https://github.com/paulhibbing/SBprofiles
}
