% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile.R
\name{get_profile}
\alias{get_profile}
\alias{get_profile.bout}
\alias{get_profile.data.frame}
\title{Implement the sedentary profiles}
\usage{
get_profile(object, ...)

\method{get_profile}{bout}(object, method = c("both", "decisionTree", "randomForest"), ...)

\method{get_profile}{data.frame}(
  object,
  method = c("both", "decisionTree", "randomForest"),
  id = NULL,
  counts,
  sb = 100,
  ...
)
}
\arguments{
\item{object}{input (either data frame or output from
\code{\link{sb_bout_dist}})}

\item{...}{further arguments passed to \code{\link{sb_bout_dist}} (used only
in data frame method)}

\item{method}{character. The model(s) to use for the prediction}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior}

\item{sb}{integer. The cut point to use for classifying sedentary behavior}
}
\value{
one or more profile assignments
}
\description{
Implement the sedentary profiles
}
\examples{
data(example_data, package = "SBprofiles")
get_profile(example_data, counts = "PAXINTEN")
get_profile(
  example_data, counts = "PAXINTEN", method = "randomForest"
)
get_profile(example_data, id = "PAXDAY", counts = "PAXINTEN")
get_profile(
  example_data, id = "PAXDAY",
  counts = "PAXINTEN", method = "decisionTree"
)
}
