% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers_and_formatters.R, R/get_bouts.R,
%   R/zzz.R
\name{id_verify}
\alias{id_verify}
\alias{counts_verify}
\alias{id_bind}
\alias{sb_bout_dist_default}
\alias{sb_bout_dist_df}
\alias{internal_functions}
\title{Functions used internally in the package}
\usage{
id_verify(object, id)

counts_verify(object, counts)

id_bind(result, id)

sb_bout_dist_default(is_sb, is_wear, min_bout, valid_indices, probs)

sb_bout_dist_df(
  df,
  counts,
  wear,
  id,
  sb,
  min_bout,
  valid_indices,
  probs,
  simplify
)
}
\arguments{
\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior}

\item{result}{an output data frame that may need id-based formatting}

\item{is_sb}{logical vector reflecting minute-by-minute classifications
(\code{TRUE} for sedentary behavior and \code{FALSE} for non-sedentary
behavior)}

\item{is_wear}{logical vector reflecting minute-by-minute wear
classifications (\code{TRUE} for wearing and \code{FALSE} for not wearing)}

\item{min_bout}{integer scalar. Threshold for a run to qualify as a bout}

\item{valid_indices}{integer vector (optional) specifying which indices of
\code{is_sb} and {is_wear} correspond to valid measurement days (e.g. with
10+ hours of wear time on 4+ days)}

\item{probs}{numeric vector. Percentiles to calculate (all values must be
between 0 and 1)}

\item{df}{data frame input (for vectorized operations)}

\item{wear}{character scalar [optional]. Column name of the variable to use
for determining wear time (logical vector with \code{TRUE} for wear time
minutes). If no value is provided, \code{\link{nhanes_wear}} is invoked on
the \code{counts} column}

\item{sb}{integer. The cut point to use for classifying sedentary behavior}

\item{simplify}{logical. If passing a data frame, setting to \code{FALSE}
will return separate bout objects that can be passed into
\code{\link{get_profile}}). Setting to \code{TRUE} will return an
aggregated data frame through \code{rbind} and \code{\link{id_bind}}.}
}
\description{
Functions used internally in the package
}
\keyword{internal}
